Basic rank system implemented + material issue
When a player dies they get added to the ranking list and then check if their is only one player left. When there is only one player left I stop all the coroutines and disallow the player input and display the winners name.
Current issue is when the players can select their color the material doesn't get saved in the player info SO